# apis/api.yaml
openapi: 3.0.0
info:
  title: Canvara Backend API
  version: 0.11.0
servers: []
...
paths:
...
  '/posts/{post_id}/applications':
    get:
      summary: List all applications for a post
      tags:
        - applications
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
      operationId: list_applications_by_post
      security:
        - AWSCognitoCanvaraPrototype: []
      parameters:
        - schema:
            type: string
            enum:
              - all
              - active
              - selected
              - passed
            default: all
          in: query
          name: filter
          description: Filter for the results.
    post:
      summary: Add a new application for a post
      operationId: create_application
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
      tags:
        - applications
    parameters:
      - schema:
          type: string
        name: post_id
        in: path
        required: true
...
  /applications:
    get:
      summary: List all applications
      tags:
        - applications
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
      operationId: list_applications_by_applicant
      security:
        - AWSCognitoCanvaraPrototype: []
  '/applications/{application_id}':
    parameters:
      - schema:
          type: string
        name: application_id
        in: path
        required: true
    get:
      summary: Get details of an application
      tags:
        - applications
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
      operationId: get_application
      security:
        - AWSCognitoCanvaraPrototype: []
    put:
      summary: Update details of an application
      operationId: update_application
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                status:
                  type: string
                  enum:
                    - active_read
                    - selected
                    - passed
      tags:
        - applications
    delete:
      summary: Delete an application
      operationId: delete_application
      responses:
        '204':
          description: No Content
      security:
        - AWSCognitoCanvaraPrototype: []
      tags:
        - applications
  '/applications/{application_id}/video':
    parameters:
      - schema:
          type: string
        name: application_id
        in: path
        required: true
        description: Unique identifier of the application.
    put:
      summary: Create request for an application video upload
      operationId: upload_application_video
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUploadResponse'
      tags:
        - applications
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUploadRequest'
            examples:
              Example:
                value:
                  filename: my_video.mp4
                  content_type: video/mp4
  '/applications/{application_id}/video/{upload_id}':
    parameters:
      - schema:
          type: string
        name: application_id
        in: path
        required: true
        description: Unique identifier of the application.
      - schema:
          type: string
        name: upload_id
        in: path
        required: true
        description: Unique identifier of the file upload.
    put:
      summary: Finalize application video upload
      operationId: finalize_application_video_upload
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
      tags:
        - applications
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: Status of the file upload.
                  enum:
                    - uploaded
              required:
                - status
    delete:
      summary: Delete an application video
      operationId: delete_application_video
      responses:
        '204':
          description: No Content
      tags:
        - applications
      security:
        - AWSCognitoCanvaraPrototype: []
...
components:
  schemas:
...
    Application:
      title: Application
      type: object
      properties:
        application_id:
          type: string
          description: Unique identifier for this application.
        post_id:
          type: string
          description: Unique identifier for the post.
        post:
          type: object
          description: Information about post.
          properties:
            post_id:
              type: string
              description: Unique identifier for this post.
            owner:
              type: object
              description: Information about the post owner.
              properties:
                user_id:
                  type: string
                  description: Unique identifier for the user.
                name:
                  type: string
                  description: Name of the user.
                profile_picture_url:
                  type: string
                  description: The URL of user's profile picture.
                title:
                  type: string
                  description: Title of the user.
                pronoun:
                  type: string
                  description: Pronoun.
                location:
                  type: string
                  description: Location of the user.
                department:
                  type: string
                  description: Department of the user.
            name:
              type: string
              description: Name of the post.
            description:
              type: string
              description: Description of this post.
            hashtags:
              type: array
              items:
                type: string
              description: List of hastags or keywords.
            status:
              type: string
              description: Current status of this post.
            created_at:
              type: string
              description: 'Date and time when this post was created, in ISO 8601 format.'
            last_updated_at:
              type: string
              description: 'Date and time when this post was last updated, in ISO 8601 format.'
        applicant_id:
          type: string
          description: Unique identifier for the user applying for the post.
        applicant:
          type: object
          description: Information about the applicant.
          properties:
            user_id:
              type: string
              description: Unique identifier for the user.
            name:
              type: string
              description: Name of the user.
            profile_picture_url:
              type: string
              description: The URL of user's profile picture.
            title:
              type: string
              description: Title of the user.
            pronoun:
              type: string
              description: Pronoun.
            location:
              type: string
              description: Location of the user.
            department:
              type: string
              description: Department of the user.
            email:
              type: string
              description: Email of the user.
            phone_number:
              type: string
              description: Phone number of the user.
            slack_teams_messaging_id:
              type: string
              description: Messaging handle of the user.
        description:
          type: string
          description: Additional information for this application.
        video_url:
          type: string
          description: The URL of video for this application.
        description_video:
          $ref: '#/components/schemas/UserUpload'
        matched_skills:
          type: array
          description: Post's skills that match the user's skills.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        unmatched_skills:
          type: array
          description: Post's skills that don't match the user's skills.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        status:
          type: string
          description: Current status of this application.
        created_at:
          type: string
          description: 'Date and time when this application was created, in ISO 8601 format.'
        last_updated_at:
          type: string
          description: 'Date and time when this application was last updated, in ISO 8601 format.'
...
    Skill:
      title: Skill
      type: object
      properties:
        skill_id:
          type: string
        name:
          type: string
    SkillWithLevel:
      title: SkillWithLevel
      allOf:
        - $ref: '#/components/schemas/Skill'
        - type: object
          properties:
            level:
              type: integer
...
    UserUploadRequest:
      title: UserUploadRequest
      type: object
      properties:
        filename:
          type: string
          description: Name of the file being uploaded.
        content_type:
          type: string
          description: Content type of the file being uploaded.
      required:
        - filename
        - content_type
    UserUploadResponse:
      title: UserUploadResponse
      type: object
      properties:
        upload_id:
          type: string
          description: Unique identifier of this upload.
        url:
          type: string
          description: Url where the file should be uploaded to.
      required:
        - upload_id
        - url
    UserUpload:
      title: UserUpload
      type: object
      properties:
        upload_id:
          type: string
          description: Unique identifier of this upload.
        url:
          type: string
          description: Url for the uploaded file.
        content_type:
          type: string
          description: Content type of the uploaded file.
      required:
        - url
        - content_type
...
  securitySchemes:
    AWSCognitoCanvaraPrototype:
      type: http
      scheme: bearer